% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiBit.R
\name{bibit}
\alias{bibit}
\title{bibit algorithm}
\usage{
bibit(matrix_binary)
}
\arguments{
\item{matrix_binary}{a binary matrix with values 0 and 1}
}
\value{
list of biclusters (row and columns indexes)
}
\description{
Executing bibit algorithm
}
