% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimax.R
\name{bimax}
\alias{bimax}
\title{BiMax algorithm}
\usage{
bimax(E)
}
\arguments{
\item{E}{a binary matrix with values 0 and 1}
}
\value{
a vector of biclusters (G,C), G bicluster'rows, C bicluster'columns
}
\description{
Find all inclusion-maximal biclusters in a binary matrix using the BiMax algorithm
}
