% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimax.R
\name{conquer}
\alias{conquer}
\title{Conquer function of BiMax algorithm}
\usage{
conquer(E, submat, Z)
}
\arguments{
\item{E}{a binary matrix with values 0 and 1}

\item{submat}{a list [G,C], the vector G corresponds to the submatrix'rows, C corresponds to the submatrix'columns}

\item{Z}{list of vector of columns CV}
}
\value{
a vector of biclusters (G,C), G bicluster'rows, C bicluster'columns
}
\description{
Find all inclusion-maximal biclusters in a binary submatrix using the BiMax algorithm
}
