% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimax.R
\name{divide}
\alias{divide}
\title{Divide function of BiMax algorithm}
\usage{
divide(E, submat, Z)
}
\arguments{
\item{submat}{a list [G,C], the vector G corresponds to the submatrix'rows, C corresponds to the submatrix'columns}

\item{Z}{list of vector of columns CV}
}
\value{
a list with GU, GV, GW, CU, CV
}
\description{
Find U and V submatrices
}
