% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimax.R
\name{reduce}
\alias{reduce}
\title{Reduce function of BiMax algorithm}
\usage{
reduce(E, submat, Z)
}
\arguments{
\item{submat}{a list [G,C], the vector G corresponds to the submatrix'rows, C corresponds to the submatrix'columns}

\item{Z}{list of vector of columns CV}
}
\value{
Gprim variable used in divide function
}
\description{
Index of rows that have at least value 1 among the C columns + Z condition
}
